requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.2
celery==5.5.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dj-database-url==3.0.1
Django==5.0.4
django-ratelimit==4.1.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
gunicorn==23.0.0
idna==3.7
inflection==0.5.1
kombu==5.5.4
packaging==25.0
pillow==10.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
requests==2.31.0
six==1.17.0
sqlparse==0.5.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
